#2119 Support for the OPC UA protocol using the PLC4X library: - Refactoring: interface PollingService (Using the PointValueTime type instead of the generic Object), OpcUaDataType(Using the MangoValue type instead of the generic Object); - Refactoring class OpcUaConverterUtils change name methods: toLongFromDouble, toByteFromDouble, toIntFromDouble, toFloatFromString, toDoubleFromString, toBooleanFromString; - Removed a redundant class (In MangoValue class there is a method numberValue); - Corrected method validation in OpcUaBaseDataType;
