+++ src/org/scada_lts/ds/polling/protocol/opcua/client/impl/OpcUaService.java
@@ -54,7 +54,7 @@ public class OpcUaService implements IOpcUaService {
             OpcUaClient opcUaClient = OpcUaClientFactory.createClient(dataSource);
             this.client = opcUaClient.connect().get(dataSource.getDefaultTimeout(), TimeUnit.MILLISECONDS);
             this.dataTypeTree = DataTypeTreeBuilder.build(opcUaClient);
-        } catch (Exception ex) {
+        } catch (Throwable ex) {
             LOG.warn(LoggingUtils.exceptionInfo(ex), ex);
             this.terminate();
             throw new PollingServiceException(ex.getMessage(), ex);
@@ -87,7 +87,7 @@ public class OpcUaService implements IOpcUaService {
         try {
             UaClient client = getClient();
             readResponse = OpcUaUtils.sendRead(client, nodeIds);
-        } catch (Exception ex) {
+        } catch (Throwable ex) {
             LOG.warn(LoggingUtils.exceptionInfo(ex), ex);
             throw new PollingServiceException(ex.getMessage(), ex);
         } finally {}
@@ -127,12 +127,13 @@ public class OpcUaService implements IOpcUaService {
     }
 
     @Override
-    public void write(DataPointVO dataPoint, Object value) throws PollingServiceException {
+    public void write(DataPointVO dataPoint, PointValueTime pointValueTime) throws PollingServiceException {
         OpcUaPointLocatorVO pointLocator = dataPoint.getPointLocator();
-        if(value == null) {
+        if(pointValueTime == null || pointValueTime.getValue() == null || pointValueTime.getValue().toValue() == null) {
             throw new PollingServiceException(getMessage("Write", pointLocator, null, "Value is null!"));
         }
-        if(value instanceof ImageValue) {
+        MangoValue value = pointValueTime.getValue();
+        if(value.toValue() instanceof ImageValue) {
             throw new PollingServiceException(getMessage("Write",pointLocator, value, value.getClass().getName() + " is not supported!"));
         }
         if(!pointLocator.getOpcDataType().validate(value)) {
@@ -141,7 +142,7 @@ public class OpcUaService implements IOpcUaService {
         Object valueToSend = null;
         try {
             valueToSend = pointLocator.getOpcDataType().convertToWrite(value);
-        } catch (Exception e) {
+        } catch (Throwable e) {
             throw new PollingServiceException(e.getMessage(), e);
         }
         try {
@@ -161,9 +162,13 @@ public class OpcUaService implements IOpcUaService {
                 }
             }
 
-        } catch (Exception ex) {
+        } catch (Throwable ex) {
             LOG.warn(LoggingUtils.exceptionInfo(ex), ex);
-            throw new PollingServiceException(ex.getMessage(), ex);
+            if(ex instanceof PollingServiceException) {
+                throw (PollingServiceException)ex;
+            } else {
+                throw new PollingServiceException(ex.getMessage(), ex);
+            }
         } finally {
         }
     }
@@ -176,7 +181,7 @@ public class OpcUaService implements IOpcUaService {
                 throw new IllegalStateException("No connected!");
             }
             List<DataValue> result = sendReadServerStateAndTime(client);
-        } catch (Exception ex) {
+        } catch (Throwable ex) {
             LOG.warn(LoggingUtils.exceptionInfo(ex), ex);
             throw new PollingServiceException(ex.getMessage(), ex);
         } finally {
@@ -211,7 +216,7 @@ public class OpcUaService implements IOpcUaService {
                 List<OpcUaPointLocatorVO> items = new ArrayList<>(result);
                 items.sort(comparator);
                 return items;
-            } catch (Exception ex) {
+            } catch (Throwable ex) {
                 LOG.warn(LoggingUtils.exceptionInfo(ex), ex);
                 return Collections.emptyList();
             } finally {
@@ -245,7 +250,7 @@ public class OpcUaService implements IOpcUaService {
             if(readResponse.getResults()[0].getStatusCode() == null)
                 return false;
             return readResponse.getResults()[0].getStatusCode().isGood();
-        } catch (Exception ex) {
+        } catch (Throwable ex) {
             LOG.warn(LoggingUtils.exceptionInfo(ex), ex);
             return false;
         } finally {
@@ -296,8 +301,8 @@ public class OpcUaService implements IOpcUaService {
                 Object value = Array.get(valueRaw, 0);
                 mangoValue = pointLocator.getOpcDataType().convertToRead(value);
             }
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        } catch (Throwable e) {
+            throw new PollingServiceException(e.getMessage(), e);
         }
         return new PointValueTime(mangoValue, time);
     }
