+++ test/org/scada_lts/ds/polling/protocol/opcua/vo/OpcUaDataTypeTestsSuite.java
@@ -6,8 +6,8 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
         IsPossibleSettableOpcUaDataTypeTest.class,
+        ValidateOpcUaBasicDataTypeTest.class,
         ValidateOpcUaDataTypeTest.class,
-        ValidateOpcUaDataType1Test.class,
 })
 public class OpcUaDataTypeTestsSuite {
 }
